requirements.txt
Django==4.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
pika==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0



